chore: pass country to growthbook on login using cloudflare check
