+++ packages/utils/src/index.ts
@@ -9,6 +9,7 @@ export { default as groupTransactionsByDay } from './groupTransactionsByDay';
 export { default as getTruncatedString } from './getTruncatedString';
 export { default as unFormatLocaleString } from './unFormatLocaleString';
 export { default as getLoginHistoryFormattedData } from './getLoginHistoryFormattedData';
+export { default as getCountry } from './getCountry';
 export * from './parse-url';
 export * from './moment';
 export * from './files';
